removed integration tests and old adapter
